+++ include/GRMustacheVersion.h
@@ -33,12 +33,12 @@
  
  @since v1.0
  */
-#define GRMUSTACHE_MINOR_VERSION 11
+#define GRMUSTACHE_MINOR_VERSION 12
 
 /**
  The patch-level component of GRMustache version
  
  @since v1.0
  */
-#define GRMUSTACHE_PATCH_VERSION 2
+#define GRMUSTACHE_PATCH_VERSION 0